mini_app/backend/requirements.txt
# SMM Mini App Backend - Requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# ASGI
starlette==0.35.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0

# JWT Auth
python-jose[cryptography]==3.3.0

# Environment
python-dotenv==1.0.0

# Database (SQLite - built-in)
# No additional package needed

# Optional: Production
gunicorn==21.2.0
